feat: 선호 플랫폼 기반 리다이렉션, GET /api/links/{shortCode}/platforms

+++ src/main/java/com/playona/api/domain/user/dto/PlatformPreferenceRequest.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 @Getter
 @Setter
 public class PlatformPreferenceRequest {
-    private Long platformId;
+    private Integer platformId;
     private Integer priority;
 }
