+++ src/main/java/com/playona/api/domain/user/dto/PlatformPreferenceResponse.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class PlatformPreferenceResponse {
-    private Long platformId;
+    private Integer platformId;
     private String platformName;
     private Integer priority;
 }
